\docType{package}
\name{variosig-package}
\alias{variosig-package}
\title{Spatial Dependence Based on Empirical Variograms}
\description{
Apply Monte Carlo permutation to compute pointwise variogram envelopes, and check spatial dependence using permutation test adjusted for multiple testing.
}
\details{
\tabular{ll}{
Package: \tab variosig\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2018-11-08\cr
License: \tab GPL (>= 3) \cr
LazyLoad: \tab yes\cr
}}

\author{
Craig Wang \email{craig.wang@uzh.ch} \cr
}
\keyword{package}

