\docType{package}
\name{variosig-package}
\alias{variosig-package}
\title{Testing Spatial Dependence Using Empirical Variogram}
\description{
Applying Monte Carlo permutation to generate pointwise variogram envelope and checking for spatial dependence at different scales using permutation test. Empirical Brown's method and Fisher's method are used to compute overall p-value for hypothesis test.
}
\details{
\tabular{ll}{
Package: \tab variosig\cr
Type: \tab Package\cr
Version: \tab 0.3-1\cr
Date: \tab 2020-02-08\cr
License: \tab GPL (>= 3) \cr
LazyLoad: \tab yes\cr
}}

\author{
Craig Wang \email{craig.wang@uzh.ch} \cr
}
\keyword{package}

